splitting the tests back out for speed
